Move CLI into a submodule
